requirements.txt
# === Core ===
streamlit>=1.40.0
pydantic-settings>=2.0.0

# === Code Parsing ===
tree-sitter>=0.23.0
tree-sitter-python>=0.23.0
tree-sitter-javascript>=0.23.0
tree-sitter-java>=0.23.0
tree-sitter-typescript>=0.23.0

# === Embedding (UniXcoder) ===
transformers>=4.40.0
torch>=2.0.0

# === Vector DB ===
qdrant-client>=1.12.0

# === Git ===
gitpython>=3.1.40

# === File Filtering ===
pathspec>=0.12.0

# === LLM (Ollama) ===
ollama>=0.4.0
